* moved models to model folder * changed buggy type in Login.xml * expect OAuth2Interface in AccessTokenAuthentication and RefreshTokenData * removed null check for AccessTokenAuthentication and RefreshTokenData
