+++ resources/serializer/OAuth2.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <serializer>
-  <class name="Mi\VideoManagerPro\SDK\Common\Token\OAuth2" exclusion-policy="ALL">
+  <class name="Mi\VideoManagerPro\SDK\Model\OAuth2" exclusion-policy="ALL">
     <property name="accessToken" serialized-name="accessToken" type="string" since-version="1.0" expose="true" />
     <property name="refreshToken" serialized-name="refreshToken" type="string" since-version="1.0" expose="true" />
   </class>
