+++ src/Common/Subscriber/RefreshTokenData.php
@@ -6,6 +6,7 @@ use GuzzleHttp\Command\Event\PreparedEvent;
 use GuzzleHttp\Command\Guzzle\Description;
 use GuzzleHttp\Event\SubscriberInterface;
 use GuzzleHttp\Stream\Stream;
+use Mi\VideoManagerPro\SDK\Common\Token\OAuth2Interface;
 
 /**
  * @author Alexander Miehe <alexander.miehe@movingimage.com>
@@ -17,9 +18,9 @@ class RefreshTokenData implements SubscriberInterface
 
     /**
      * @param Description    $description
-     * @param string $refreshToken
+     * @param OAuth2Interface $refreshToken
      */
-    public function __construct(Description $description, $refreshToken = null)
+    public function __construct(Description $description, OAuth2Interface $refreshToken)
     {
         $this->description = $description;
         $this->refreshToken = $refreshToken;
@@ -35,15 +36,11 @@ class RefreshTokenData implements SubscriberInterface
 
     public function onPrepared(PreparedEvent $event)
     {
-        if ($this->refreshToken === null) {
-            return;
-        }
-
         $command   = $event->getCommand();
         $operation = $this->description->getOperation($command->getName());
 
         if ($operation->getData('refresh-token-data') === true) {
-            $event->getRequest()->setBody(Stream::factory(json_encode(['refreshToken' => $this->refreshToken])));
+            $event->getRequest()->setBody(Stream::factory(json_encode(['refreshToken' => $this->refreshToken->getRefreshToken()])));
         }
     }
 }
