+++ src/Model/VideoManager.php
@@ -4,6 +4,8 @@ namespace Mi\VideoManagerPro\SDK\Model;
 
 /**
  * @author Steve Reichenbach <steve.reichenbach@movingimage.com>
+ *
+ * @codeCoverageIgnore
  */
 class VideoManager
 {
